fix compiler warnings

+++ include/object.h
@@ -505,7 +505,7 @@ extern object *Player_obj;	// Which object is the player. Has to be valid.
 // Use this instead of "objp - Objects" to get an object number
 // given it's pointer.  This way, we can replace it will a macro
 // to check that the pointer is valid for debugging.
-#define OBJ_INDEX(objp) (objp-Objects)
+#define OBJ_INDEX(objp) (static_cast<int>(objp-Objects))
 
 /*
  *		FUNCTIONS
